[FEAT/#92] 공고 상세 페이지 공고 아이디에 따라 조회

+++ data/src/main/java/com/terning/data/datasource/InternDataSource.kt
@@ -4,5 +4,5 @@ import com.terning.data.dto.BaseResponse
 import com.terning.data.dto.response.InternResponseDto
 
 interface InternDataSource {
-    suspend fun getInternInfo(id: Int): BaseResponse<InternResponseDto>
+    suspend fun getInternInfo(id: Long): BaseResponse<InternResponseDto>
 }
