+++ domain/src/main/java/com/terning/domain/repository/InternRepository.kt
@@ -3,5 +3,5 @@ package com.terning.domain.repository
 import com.terning.domain.entity.response.InternInfoModel
 
 interface InternRepository {
-    suspend fun getInternInfo(id: Int): Result<InternInfoModel>
+    suspend fun getInternInfo(id: Long): Result<InternInfoModel>
 }
